% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_magnet_har.R
\name{magnet_base}
\alias{magnet_base}
\title{Load magnet base year data in long format in R}
\usage{
magnet_base(indicators, regions, scenarios, base_year, file_path, file_type)
}
\description{
Load magnet base year data in long format in R
}
