% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_magnet_har.R
\name{magnet_indicator}
\alias{magnet_indicator}
\title{Load specific indicators from MAGNET in long format in R}
\usage{
magnet_indicator(indicator, scenarios, periods, base_year, path_project)
}
\description{
Load specific indicators from MAGNET in long format in R
}
