% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_magnet_har.R
\name{magnet_scenario}
\alias{magnet_scenario}
\title{Load magnet results in long format in R

\code{magnet_scenario} load results from the general computable general equilibrium model 'MAGNET' in long format}
\usage{
magnet_scenario(
  indicators,
  scenarios,
  periods,
  file_path,
  file_type,
  file_suffix
)
}
\arguments{
\item{indicators}{The indicator name quotation marks or a vector of the indicator names, indicator name(s) need to correspond with the Header names in the .har file}

\item{scenarios}{The scenario name in quotation marks or a vector of scenario names}

\item{periods}{The period name in quotation marks or a vector of the period names}

\item{file_path}{The file path where the 'MAGNET' results in e.g. .har or .sol files are stored}

\item{file_type}{This corresponds with from where the results are extracted, like from a .har or .sol file with the name "Update", "Solution" or "update_view"}

\item{file_suffix}{The suffix of the data file e.g. har or sol}

\item{regions}{The region name quotation marks or a vector of the region names, region name(s) corresponds with abbreviations used in magnet}
}
\description{
Load magnet results in long format in R

\code{magnet_scenario} load results from the general computable general equilibrium model 'MAGNET' in long format
}
\examples{

\dontrun{

raw_path <- "W:/WECR/magnet_data9/PCSL/4_magnet"/"Updates"
scenarios <-c("Base_PCSL_ProdPath_YouthBranch" , "Base_PCSL_ProdPath", "Base_PCSL_Extpath")
periods <- c("2014-2020", "2020-2025", "2025-2030", "2030-2035", "2035-2040", "2040-2045")
country <- c("Ken")
magnet_scenario(scenarios, periods, "POP", country,"update", dataUpdatesPath, "har")
}

}
