% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnetr-package.R
\docType{package}
\name{magnetr}
\alias{magnetr}
\title{magnetr: A package to load results from the global computable general equilibrium model 'MAGNET'}
\description{
ADD description
}
\section{magnetr functions}{

extract_MAGNET_d1()
....
}

